scripts/Document-scraping-Tool/requirements.txt
requests
colorama
argparse
bs4
PyExifTool